+++ apps/client/src/app/components/admin-market-data/admin-market-data.component.ts
@@ -13,7 +13,6 @@ import {
   User
 } from '@ghostfolio/common/interfaces';
 import { hasPermission, permissions } from '@ghostfolio/common/permissions';
-import { GfSymbolPipe } from '@ghostfolio/common/pipes';
 import { GfActivitiesFilterComponent } from '@ghostfolio/ui/activities-filter';
 import { GfFabComponent } from '@ghostfolio/ui/fab';
 import { translate } from '@ghostfolio/ui/i18n';
@@ -83,7 +82,6 @@ import { CreateAssetProfileDialogParams } from './create-asset-profile-dialog/in
     GfActivitiesFilterComponent,
     GfFabComponent,
     GfPremiumIndicatorComponent,
-    GfSymbolPipe,
     GfValueComponent,
     IonIcon,
     MatButtonModule,
@@ -410,7 +408,8 @@ export class GfAdminMarketDataComponent implements AfterViewInit, OnInit {
 
         const dialogRef = this.dialog.open<
           GfAssetProfileDialogComponent,
-          AssetProfileDialogParams
+          AssetProfileDialogParams,
+          AssetProfileIdentifier
         >(GfAssetProfileDialogComponent, {
           autoFocus: false,
           data: {
@@ -428,15 +427,13 @@ export class GfAdminMarketDataComponent implements AfterViewInit, OnInit {
         dialogRef
           .afterClosed()
           .pipe(takeUntilDestroyed(this.destroyRef))
-          .subscribe(
-            (newAssetProfileIdentifier: AssetProfileIdentifier | undefined) => {
-              if (newAssetProfileIdentifier) {
-                this.onOpenAssetProfileDialog(newAssetProfileIdentifier);
-              } else {
-                this.router.navigate(['.'], { relativeTo: this.route });
-              }
+          .subscribe((newAssetProfileIdentifier) => {
+            if (newAssetProfileIdentifier) {
+              this.onOpenAssetProfileDialog(newAssetProfileIdentifier);
+            } else {
+              this.router.navigate(['.'], { relativeTo: this.route });
             }
-          );
+          });
       });
   }