+++ libs/common/src/lib/pipes/index.ts
@@ -1,3 +1,0 @@
-import { GfSymbolPipe } from './symbol.pipe';
-
-export { GfSymbolPipe };